Move OnSeekChangeListener logic back to VideoControllerView
